+++ src/matbuild.c
@@ -309,7 +309,7 @@ s32 loadMat(nfs_sparse_mat_t *M, char *colName)
   for (i=0; i<M->numDenseBlocks; i++) {
     if (!(M->denseBlocks[i] = (u64 *)lxcalloc((M->numCols+1)*sizeof(u64),0))) {
       fprintf(stderr, "loadMat() Error allocating %" PRIu32 " bytes for the QCB entries!\n",
-              (M->numCols+1)*sizeof(u64));
+              (u32)((M->numCols+1)*sizeof(u64)) );
       free(M->cIndex); free(M->cEntry); fclose(fp); return -1;
     }
   }
@@ -375,12 +375,12 @@ rel_list *getRelList(multi_file_t *prelF, int index)
   /* Now allocate for the relations. */
   if (!(RL->relData = (s32 *)lxmalloc(RL->maxDataSize * sizeof(s32),0))) {
     fprintf(stderr, "Error allocating %" PRIu32 "MB for reading relation list!\n",
-            RL->maxDataSize * sizeof(s32)/1048576);
+            (u32)(RL->maxDataSize * sizeof(s32)/1048576) );
     free(RL); return NULL;
   }
   if (!(RL->relIndex = (s32 *)lxmalloc(RL->maxRels * sizeof(s32),0))) {
     fprintf(stderr, "Error allocating %" PRIu32 "MB for relation pointers!\n", 
-            RL->maxRels * sizeof(s32)/1048576);
+            (u32)(RL->maxRels * sizeof(s32)/1048576) );
     free(RL->relData); free(RL);
     return NULL;
   }
@@ -1116,7 +1116,7 @@ int allocateRL(multi_file_t *prelF, rel_list *RL)
   RL->maxDataSize = 1000 + maxSize/sizeof(s32);
   if (!(RL->relData = (s32 *)lxmalloc(RL->maxDataSize * sizeof(s32),0))) {
     fprintf(stderr, "Error allocating %" PRIu32 "MB for processed relation files!\n",
-            RL->maxDataSize * sizeof(s32)/1048576);
+            (u32)(RL->maxDataSize * sizeof(s32)/1048576) );
     fprintf(stderr, "Try decreasing DEFAULT_MAX_FILESIZE and re-running.\n");
     exit(-1);
   }
@@ -1124,7 +1124,7 @@ int allocateRL(multi_file_t *prelF, rel_list *RL)
   RL->maxRels = (u32)RL->maxDataSize/20;
   if (!(RL->relIndex = (s32 *)lxmalloc(RL->maxRels * sizeof(s32),0))) {
     fprintf(stderr, "Error allocating %" PRIu32 "MB for relation pointers!\n",
-            RL->maxRels * sizeof(s32)/1048756);
+            (u32)(RL->maxRels * sizeof(s32)/1048756) );
     free(RL->relData);
     exit(-1);
   }
